refactor: move export logic to orchestrator for consistency
Make multi-benchmark flow consistent with single-benchmark flow where
SystemController handles execution + export internally.

Changes:
- Add execute_and_export() method to orchestrator (main entry point)
- Add _aggregate_and_export() to orchestrate aggregation + export
- Move _export_confidence_aggregate() from cli_runner to orchestrator
- Move _export_sweep_aggregates() from cli_runner to orchestrator
- Remove _print_aggregate_summary() (no longer needed)
- Simplify cli_runner to just call orchestrator.execute_and_export()

Benefits:
- Consistent with _run_single_benchmark pattern
- Orchestrator owns execution, aggregation, AND export (single responsibility)
- No hardcoded paths in CLI layer
- All path logic centralized in orchestrator
- CLI runner only handles user-facing messages and logging

Pattern:
  Single benchmark: CLI → SystemController (execute + export)
  Multi benchmark:  CLI → Orchestrator (execute + aggregate + export)
